refactor(ethereum): consolidate log parsing file
